+++ tests/SubjectPreviewTest.php
@@ -131,21 +131,4 @@ class SubjectPreviewTest extends TestCase
             ],
         ];
     }
-
-    public function testGetEmailClient(): void
-    {
-        $checkValue = sha1(uniqid());
-
-        $subjectPreview = new SubjectPreview();
-        $subjectPreview->setSubject($checkValue);
-        $subjectPreview->setBody('');
-        $subjectPreview->setSender('');
-
-        $emailClient = $subjectPreview->getEmailClient('ol2003');
-
-        self::assertEquals('ol2003', $emailClient->getSlug());
-
-        // Test EmailClient is using the given subject
-        self::assertStringContainsString($checkValue, $emailClient->getUrl(false));
-    }
 }